feat(image): Add initial Nix config to build LXC image
Adds Nix configuration that produces an output with two tarballs which
can be cleanly imported as an LXC image like so:

    lxc image import result/nixos-system-x86_64-linux.tar.gz result/nixos-system-x86_64-linux.tar.g z.root --alias nixini

Note that the init script of this container is located at `/init`, but
expected by LXC to be at `/sbin/init`. Containers must thus be
configured with `lxc config set $container raw.lxc lxc.init.cmd=/init`
to function.